Enable additional low-noise golangci-lint linters
Enable a set of correctness guards and small cleanup/modernization
linters that are either already clean on the tree or require only
mechanical fixes:

  errname, usestdlibvars, nilerr, bodyclose, durationcheck, makezero,
  reassign, predeclared, wastedassign, unconvert, whitespace, intrange,
  copyloopvar, misspell

errname reinforces the existing sentinel-error naming convention; the
correctness linters (nilerr, bodyclose, durationcheck, makezero,
reassign, predeclared, wastedassign) are clean today and guard against
future regressions.

Fixes applied for the findings this surfaced:

  - errname: rename noIdentityError -> errNoIdentity, fakeTimeoutErr ->
    fakeTimeoutError, and the osStatus error type -> osStatusError (the
    former osStatusError constructor becomes newOSStatusError).
  - unconvert: drop redundant conversions of already-typed constants
    (jceks entry tags, tls cipher suites, proxyproto command/protocol).
  - usestdlibvars: use http.Method* constants; add the net/http import
    where the rewrite required it.
  - intrange/copyloopvar/whitespace/misspell: mechanical modernizations
    and cleanups applied via golangci-lint --fix.

+++ tls_test.go
@@ -395,8 +395,8 @@ func TestResolveCipherSuites(t *testing.T) {
 
 	suites, err = resolveCipherSuites("AES,UNSAFE-AZURE", true)
 	assert.Nil(t, err, "mixed safe+unsafe should resolve with flag")
-	assert.Contains(t, suites, uint16(tls.TLS_AES_128_GCM_SHA256))
-	assert.Contains(t, suites, uint16(tls.TLS_ECDHE_RSA_WITH_AES_128_CBC_SHA256))
+	assert.Contains(t, suites, tls.TLS_AES_128_GCM_SHA256)
+	assert.Contains(t, suites, tls.TLS_ECDHE_RSA_WITH_AES_128_CBC_SHA256)
 }
 
 // TestValidateCipherSuitesMatchesBuildConfig pins the invariant that
